extract game output to separate file
